requirements.txt
ctransformers==0.2.5
sentence-transformers==3.0.1
pinecone-client==3.2.2
langchain==0.2.6
flask==3.0.3
python-dotenv==1.0.1
pypdf==4.2.0
langchain-pinecone==0.1.1
langchain-community==0.2.6
langchain_huggingface==0.0.3

-e .